Fix SwiftLint: extract repetitionPositionsByName, drop force-unwrap in test
